requirements.txt
# RAG完整教程 - 依赖包列表

# ============================================================================
# 核心框架
# ============================================================================
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
llama-index>=0.9.0

# ============================================================================
# LLM提供商
# ============================================================================
openai>=1.3.0
anthropic>=0.7.0

# ============================================================================
# 向量数据库
# ============================================================================
chromadb>=0.4.0
pymongo>=4.6.0

# ============================================================================
# 嵌入模型
# ============================================================================
sentence-transformers>=2.2.0
flagembedding>=1.2.0

# ============================================================================
# Web框架
# ============================================================================
streamlit>=1.29.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# ============================================================================
# 数据处理
# ============================================================================
pandas>=2.0.0
numpy>=1.26.0

# ============================================================================
# 网络请求
# ============================================================================
requests>=2.31.0
httpx>=0.25.0
arxiv>=2.0.0
wikipedia>=1.4.0

# ============================================================================
# 图数据库
# ============================================================================
networkx>=3.2.0
neo4j>=5.0.0

# ============================================================================
# 实用工具
# ============================================================================
python-dotenv>=1.0.0
tiktoken>=0.5.0
rich>=13.0.0

# ============================================================================
# Jupyter相关
# ============================================================================
jupyter>=1.0.0
jupyterlab>=4.0.0
notebook>=7.0.0
ipykernel>=6.25.0
nbconvert>=7.0.0

# ============================================================================
# 测试和评估
# ============================================================================
pytest>=7.4.0
pytest-cov>=4.1.0

# ============================================================================
# 部署相关
# ============================================================================
docker>=6.1.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ============================================================================
# 文档生成
# ============================================================================
mkdocs-material>=9.0.0
mkdocs-jupyter>=0.24.0
mkdocs-glightbox>=0.10.0

# ============================================================================
# 其他依赖
# ============================================================================
aiofiles>=23.0.0
pyyaml>=6.0.0
tenacity>=8.2.0
